[OGUI-659] Add Trigger radio buttons for env KV panel
